Create SRP using Decorators
Here I have implemented Single Responsibility using decorators in C# Language.. 
What is the use of Decorators ? 
Ans: The decorators are used to implement a functionality without making much changes to the  main method.  Its like adding sauce to pizza it does not change the pizza but adds more flavour to it. :)